[hap2] Move handle_exception to hap.py
