+++ server/hap2/hatohol/hap.py
@@ -25,6 +25,32 @@ transporter and the sub class shall not import haplib. So the functions and
 classes used in them have to be in this module.
 """
 
+import logging
+import sys
+import traceback
+
+def handle_exception(raises=(SystemExit,)):
+    """
+    Logging exception information including back trace and return
+    some information. This method is supposed to be used in 'except:' block.
+    Note that if the exception class is Signal, this method doesn't log it.
+
+    @raises
+    A sequence of exceptionclass names. If the handling exception is one of
+    it, this method just raises it again.
+
+    @return
+    A sequence of exception class and the instance of the handling exception.
+    """
+    (exctype, value, tb) = sys.exc_info()
+    if exctype in raises:
+        raise
+    if exctype is not Signal:
+        logging.error("Unexpected error: %s, %s, %s" % \
+                      (exctype, value, traceback.format_tb(tb)))
+    return exctype, value
+
+
 class Signal:
     """
     This class is supposed to raise as an exception in order to
